add additional state variables
Added the queue initialization to this script, so events can be tested with it (I have NOT removed the definition from Product.m!). 

State variables for whether a product is currently being produced are added; these are necessary to determine whether we can start making a produced (they will need to be set to false in the "product built" events).

State variables for the alternative designs have been added. Set the 2 booleans at the top of the script to true to use the alternative design policies we proposed (implemented, but i haven't had time to test them yet).