Sim.m processes events until end time and outputs results
Added a stopping condition for the simulation after X seconds (only generate new events when the event time is less than X). The program also now outputs relevant statistics to a file. Both the filename/path and stopping time can be altered as variables at the top of Sim.m