fixed 'NULL' issue with events in FEL
In trying to reduce code repetition I required returning a null pointer from the Event constructor. It turns out matlab does not have any concept of NULL/NIL/Nothing for an object type, so when i tried to set an event to NULL it broke. This is fixed by having an EventType "invalid" instead for when the constructor should not create an Event object.